Refactor BladeRack
